fix subscribe page
